Adding additional data for table coalesce
This file comes from Kaggle and purports to be a complete Pokedex of all generations of Pokemon prior to Legends: Arceus. 

https://www.kaggle.com/abcsds/pokemon

I will LEFT JOIN this table using to Coalesce function to merge the Type columns of the two tables while restricting the product table to only those Pokemon that we are interested in, the ones appearing in Arceus.